Move cell creation into counting expression
As `b.insertCell()...` in an expression it can be safely moved into for
loop's counting expression section.

function A becomes a new `for` body and once it's a FunctionDeclaration,
it is a block and therefore don't need a semicolon after.

Counter-intuitive moment: despite the fact this FD is located inside
loop body, it is hoisted and runs only once instead of being evaluated
at each loop step.